[#1072] refactor(code-generator): move list of actions in docs generation to separate file
